% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCoherence.R
\name{estimateCoherence}
\alias{estimateCoherence}
\title{Estimate the interferometric Coherence}
\usage{
estimateCoherence(
  master,
  slave,
  outputDirectory,
  fileName,
  swath = "IW1",
  polarisation = "VV,VH",
  firstBurst = 1,
  lastBurst = 9,
  aoi,
  aoiBuffer = 0,
  numCores,
  maxMemory,
  execute = FALSE
)
}
\arguments{
\item{master}{Master Sentinel-1 scene for coherence estimation.}

\item{slave}{Slave scene.}

\item{outputDirectory}{Directory for the output file.}

\item{fileName}{Name of the output file.}

\item{swath}{Swath to be processed. One of "IW1", "IW2" or "IW3".}

\item{polarisation}{Plarisations to be processed. One of "VH", "VV", or "VV,VH". Defaults to "VV,VH".}

\item{firstBurst}{First burst from the chosen swath.}

\item{lastBurst}{Last birst. Has to be higher or equal to first burst.}

\item{aoi}{sf object with area of interest to subst the output.}

\item{aoiBuffer}{Buffer size in metres for AOI.}

\item{numCores}{Number of CPUs to be used in the process.}

\item{maxMemory}{Amount of memory to be used in GB.}

\item{execute}{logical if command for esa SNAP gpt shall be executed.}
}
\description{
Coherence estimation for two corresponding Sentinel-1 SLC scenes
}
\examples{
#' # example AOI
aoi <- c(10.441054, 52.286959) \%>\%
  sf::st_point() \%>\%
  sf::st_sfc(crs = 4326)

# scenes for aoi and given criteria
scenes <-
  getScenes(
    aoi = aoi,
    bufferDist = 100,
    startDate = "2019-01-01",
    endDate = "2019-01-31",
    productType = "SLC"
  ) \%>\%
  dplyr::filter(relativeOrbitNumber == 44)

estimateCoherence(
  master = scenes$productPath[1],
  slave = scenes$productPath[2],
  outputDirectory = getwd(),
  fileName = "/test.tif",
  aoi = aoi,
  polarisation = "VV,VH",
  swath = "IW1",
  firstBurst = 1,
  lastBurst = 1,
  numCores = 8,
  maxMemory=50,
  execute = FALSE)
}
