% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScenes.R
\name{getScenes}
\alias{getScenes}
\title{Query scenes from CODE-DE EO Finder API}
\usage{
getScenes(
  aoi,
  bufferDist = NULL,
  startDate,
  endDate,
  productType,
  sensorMode = NULL,
  orbitDirection = NULL,
  relativeOrbitNumber = NULL,
  view = FALSE
)
}
\arguments{
\item{aoi}{Area of interest for the query. Has to be an sf object with polygonal or point geometry. In case of a point geometry, the buffer argument has to be set.}

\item{bufferDist}{Buffer around the AOI in meters. Defaults to 0.}

\item{startDate}{Starting date for the query of format "YYYY-MM-DD".}

\item{endDate}{End date for query.}

\item{productType}{Product type to query. One of "SLC", "GRD", "CARD-INF6", ...}

\item{sensorMode}{Sensor mode for Sentinel-1. One of "IW", "EW", ... Only considered if set.}

\item{orbitDirection}{Direction of satellite orbit. One of "ascending" or "descending". Only considered if set.}

\item{relativeOrbitNumber}{Relative orbit number. Only considered if set.}

\item{view}{logical indicating if the queried scenes should be visualized with mapview. Defaults to FALSE.}
}
\value{
sf object containing footprints and metadata for all queried scenes.
}
\description{
Returns metadata for satellite imagery that matches several input criteria
}
\examples{
# example AOI
aoi <- c(10.441054, 52.286959) \%>\%
  sf::st_point() \%>\%
  sf::st_sfc(crs = 4326)

# scenes for aoi and given criteria
scenes <-
  getScenes(
    aoi = aoi,
    startDate = "2019-01-01",
    endDate = "2019-01-15",
    productType = "SLC",
    view = TRUE
  )
}
