% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCoherence.R
\name{estimateCoherence}
\alias{estimateCoherence}
\title{Estimate the interferometric Coherence}
\usage{
estimateCoherence(
  master,
  slave,
  outputDirectory,
  fileName,
  polarisation = "VV,VH",
  swath = "IW1",
  firstBurst = 1,
  lastBurst = 9,
  aoi = NULL,
  aoiBuffer = 0,
  numCores,
  maxMemory,
  execute = FALSE,
  return = FALSE
)
}
\arguments{
\item{master}{Master Sentinel-1 scene for coherence estimation.}

\item{slave}{Slave scene.}

\item{outputDirectory}{Directory for the output file.}

\item{fileName}{Name of the output file.}

\item{polarisation}{Plarisations to be processed. One of "VH", "VV", or "VV,VH". Defaults to "VV,VH".}

\item{swath}{Swath to be processed. One of "IW1", "IW2", "IW3" or "all".}

\item{firstBurst}{First burst index from the chosen swath. Between 1 and 9. Only relevant if swath != "all".}

\item{lastBurst}{Last burst index. Has to be higher than or equal to first burst. Only relevant if swath != "all".}

\item{aoi}{sf object for the area of interest.}

\item{aoiBuffer}{Buffer around aoi in meters. Defaults to 0.}

\item{numCores}{Number of CPUs to be used in the process.}

\item{maxMemory}{Amount of memory to be used in GB.}

\item{execute}{logical if command for esa SNAP gpt shall be executed. If FALSE the commmand is printed instead.}

\item{return}{logical if processed raster or stack shall be returned.}
}
\value{

}
\description{
Coherence estimation for two corresponding Sentinel-1 SLC scenes
}
\examples{
# example AOI
aoi <- c(10.441054, 52.286959) \%>\%
  sf::st_point() \%>\%
  sf::st_sfc(crs = 4326)

# scenes for aoi and given criteria
scenes <-
  getScenes(
    aoi = aoi,
    bufferDist = 100,
    startDate = "2019-01-01",
    endDate = "2019-01-31",
    productType = "SLC"
  ) \%>\%
  dplyr::filter(relativeOrbitNumber == 44)

estimateCoherence(
  master = scenes$productPath[1],
  slave = scenes$productPath[2],
  outputDirectory = getwd(),
  fileName = "test.tif",
  polarisation = "VV,VH",
  swath = "IW1",
  firstBurst = 1,
  lastBurst = 1,
  numCores = 8,
  maxMemory=50,
  execute = FALSE)
}
